test_plan: Log warning if regexp doesn't match any files in validation.
- Make this check a warning for now, as it doesn't necessarily
mean the config is invalid, but there's a good chance someone
made an error in the regexp.

- Change the validate command to only take a single directory as
input, so we can find the root of the repo and pass this to the
validation commands. `dirmd.ReadMapping` will fail anyway if dirs
from different repos are passed in.

BUG=b:275814460
TEST=CQ

Change-Id: Iac979b5fbd13af80ae44c2526aaa6cceca1f8573
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/4385761
Commit-Queue: Navil Perez <navil@google.com>
Auto-Submit: Andrew Lamb <andrewlamb@chromium.org>
Reviewed-by: Navil Perez <navil@google.com>
Cr-Commit-Position: refs/heads/main@{#55952}