Remove qyearsley@chromium.org from recipes/recipes/wpt_import.expected/OWNERS.
This suggested change is automatically generated based on group memberships and affiliations.
If this change is unnecessary or in error, vote CR -1 and the bot will abandon it. Vote CR +1/2 to approve this change.
See the owner's recent activity for context: https://chromium-review.googlesource.com/q/qyearsley@chromium.org

To report an issue, file a bug in the Infra>Security component.

Change-Id: Ie43961acb20a442f6a804368ac446a43bdc53a41