cros_test_platform: Show correct time for Test Runner invocation
Test Runner invocations currently show the slowest step's time for all of the neighboring steps.

Ensure Test Runner invocations are closed when they complete to ensure their timing is marked correctly, as well as ensuring in the UI they are showed as complete instead of waiting on the slowest task to complete.

BUG=b:270547321
TEST=led

Change-Id: I672096f47932c3de990b2c5cc0d37f619ba43ce7
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/4409643
Reviewed-by: Derek Beckett <dbeckett@chromium.org>
Commit-Queue: Chris DeLaGarza <cdelagarza@google.com>
Reviewed-by: Jaques Clapauch <jaquesc@google.com>
Cr-Commit-Position: refs/heads/main@{#55982}