Create jammy versions of infra-continuous bots on dev and prod
The ccompute bionic image roller currently gates image rolls on
greenness of:
https://luci-milo-dev.appspot.com/p/infra/builders/ci/infra-continuous-bionic-64

To enable jammy rolls, we'd probably want a jammy version of that dev
bot. So this adds it.

This also adds its prod (ci and try) versions of that dev bot, because
they'll probably be needed eventually after bionic is EOL'ed later
this summer.

Bug: 1412588
Change-Id: Ied498f48bde38fd950a989626d40ccbca60dd93a
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/4409732
Reviewed-by: Vadim Shtayura <vadimsh@chromium.org>
Commit-Queue: Ben Pastene <bpastene@chromium.org>
Cr-Commit-Position: refs/heads/main@{#55993}