cros try retry: look for previous retries
If build B is a retry of build A, subsequent attempts to retry A
should actually retry B. When we break the linearity of retries we
can clobber builds and cause nastiness.

Possibly missing some edge cases with orchestrator reruns, will
investigate and upload a follow-on CL if needed.

BUG=b:277366983
TEST=go test, go run

Change-Id: I1d31628a9fbb2cb76ad8a43a9a451586e694f387
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/4411760
Commit-Queue: Jack Neus <jackneus@google.com>
Reviewed-by: Madeleine Hardt <hardtmad@google.com>
Cr-Commit-Position: refs/heads/main@{#56038}