manifest_doctor: Support fallback remotes
In very old branches, manifests don't have public/private annotations in
remotes, so we download the ToT manifest to get the information. However
this method doesn't work for remotes that once existed but removed in
the ToT manifest, such as "weave" remote.

This patch adds a list of "fallback" remotes to be used when the ToT
manifest lacks them. We can add fallback remotes when we remove a remote
from the ToT manifest while old manifests without remote annotations
still refer to them.

BUG=b:269194223
TEST=go test .

Change-Id: I2fc5aa94180d07cf1f2b984661d5f385fdd20338
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/4413976
Reviewed-by: Jack Neus <jackneus@google.com>
Commit-Queue: Shuhei Takahashi <nya@chromium.org>
Cr-Commit-Position: refs/heads/main@{#56090}