[recovery]: Do not call badblocks from specific actions.
The badblocks check is not needed with when invoked froma few of the
actions.

This commit adds a flag that suppresses this check when the exec is
called from these actions. The check will continue to be be done when
this exec is invoked from other actions where this is required.

Bug=b:242584223
TEST=build, go test ./...

Change-Id: Ie654a975bf38f1b6f43e4e1646375a2f94e9593f
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/4410977
Commit-Queue: Vineet Joshi <vkjoshi@google.com>
Reviewed-by: Otabek Kasimov <otabek@google.com>
Cr-Commit-Position: refs/heads/main@{#56108}