[UFS] Ignore unknown fields in UFS config loader
Bug: b/278622019

Change to google protobuf. This will allow us to use the
UnmarshalOptions.

This change if applied ignores any new fields in the config, should
prevent any crashes due to the proto mismatch if config is pushed
before the service is updated to read the config

Change-Id: I584751d0d50475ea245151eb02cd368c982a5254
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/4437612
Reviewed-by: Justin Suen <justinsuen@google.com>
Reviewed-by: Xixuan Wu <xixuan@google.com>
Commit-Queue: Anushruth Sasivehalli <anushruth@chromium.org>
Cr-Commit-Position: refs/heads/main@{#56163}