[golangbuild] split step presentation from build state
This change refactors golangbuild and makes the flow of steps clearer by
splitting out state extraction and validation, encapsulating it into a
type that represents what the build is supposed to do: a buildSpec. This
change then also defines a sourceSpec which simplifies the APIs for
fetching repositories.

Lastly, this change writes down guidelines to help keep the high-level
step presentation clear.

This refactoring is made in preparation for new complexities like
sharing toolchain snapshots and test sharding.

This change is almost a no-op. The only differences should be in the
step presentation (particularly in how commands are invoked).

Change-Id: Id23a0391bec1f5df79605e2911c57c22cafad539
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/4428755
Commit-Queue: Michael Knyszek <mknyszek@google.com>
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>
Reviewed-by: Heschi Kreinick <heschi@google.com>
Cr-Commit-Position: refs/heads/main@{#56239}