[golangbuild] set GOBUILDEXIT=1 on Windows
This causes .bat scripts on Windows to emit exit codes on failure,
a behavior we want when testing Go in a CI system. It's not on by
default because it would cause some terminals on Windows to close
the terminal window.

Change-Id: I284cd7737e1f082f96d76081fdca0a84820c1d35
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/4451166
Commit-Queue: Michael Knyszek <mknyszek@google.com>
Auto-Submit: Dmitri Shuralyov <dmitshur@google.com>
Reviewed-by: Michael Knyszek <mknyszek@google.com>
Cr-Commit-Position: refs/heads/main@{#56241}