[WIB] Add support to trigger jobs from the CL
Add support to trigger jobs from any refs. This is a temporary way to
trigger the builds from the CL. This will reduce the time taken to fix
and test the configs. Will need to be changed to running in test mode
in future (so that binaries built are uploaded to test counterparts)

Change-Id: I7647d0981506d324aeaaa379c02c701f8d0ab346