Adding support for files to dirmd
MetadataOverride is being introduced to allow users to specify
file patterns and metadata to override metadata information for
files in a directory.

MetadataOverride is processed as the directory is being read, and
is written into its own map[string]Mapping for files as defined in
mapping.proto. This is exported under a "files" key as well in the
resulting json.

This change does not include BQ, as the table will require changes
to support file information. This change also excludes ResultSink
handling, which should cross reference file_loc and files before
determining the actual metadata.

Bug: 1431748
Change-Id: I8c01b6a839e55226cfa6cddac7b7feda0e4b36d1
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/4448527
Reviewed-by: Chan Li <chanli@chromium.org>
Commit-Queue: Jeff Yoon <jeffyoon@google.com>
Cr-Commit-Position: refs/heads/main@{#56309}