vm-leaser: implement Create InstanceApi in vmlab libs
The Create method takes an instance creation request and communicates
with the VM Leaser Service backend to create a VM. Validations are added
for verifying payload.

For development purposes, environment flag and config are implemented.
Local, staging, and production environments can be specified. This is
specific to the VM Leaser backend only.

BUG=b/274054929
TEST=go test ./...; manual integration testing

Change-Id: I3b007c9e1233756bdd4c7a119aa8e1cf7e3bd031
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/4417439
Reviewed-by: Ming Kong <mingkong@google.com>
Reviewed-by: Deniz Kara <denizkara@google.com>
Commit-Queue: Justin Suen <justinsuen@google.com>
Cr-Commit-Position: refs/heads/main@{#56311}