siso: allow missing outputs configured by siso config
siso won't fail the step if step's output is missing,
if the output is not specified by build.ninja.

if the output is specified by build.ninja, it is expected
to be generated by the step, so missing the output would
be something wrong.

but if the output is not specified by build.ninja, but
added by siso config or handler, it would be safe to
ignore it as the output is not used in the build graph.

remove 'allow_missing_outputs' from step config rule.

Change-Id: Ib290932165827a24edaa135c7080bec01fc59b8c
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/4473188
Commit-Queue: Junji Watanabe <jwata@google.com>
Reviewed-by: Junji Watanabe <jwata@google.com>
Auto-Submit: Fumitoshi Ukai <ukai@google.com>
Cr-Commit-Position: refs/heads/main@{#56314}