🚝 Reduce gunicorn worker count for local development.
Having 8 workers per module leads to an explosion of 32 workers
while running the local dev server. That leads to more parallelism
and more resource consumption, which hinders development.

Ideally we could just set it to 1 for local dev, but local dev
is set up to imitate monorail-staging, so set it to 1 for both
monorail-dev and monorail-staging as well.

Change-Id: I17759ca98cd0e3089c217207c270699f9ea06564
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/4481644
Reviewed-by: Andrew Chang <andrewjc@google.com>
Commit-Queue: Andrew Chang <andrewjc@google.com>
Auto-Submit: Dave Tu <dtu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#56372}