Use the infra-system python2 for bootstrap in the luci_py builder.
This is in preparation for disabling the bootstrap in gclient hooks
by default. Other recipes that invoke infra/test.py already use the
infra-system python2.

Note that the infra-system python is not installed on all platforms
(i.e. mac-arm64) but it appears to be installed on all of the platforms
that this builder uses.

Bug: 1393867
Change-Id: Icee5003909a0dff012170b4a1e2feb68d121149f
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/4480569
Reviewed-by: Takuto Ikuta <tikuta@chromium.org>
Commit-Queue: Brian Ryner <bryner@google.com>
Cr-Commit-Position: refs/heads/main@{#56386}