🚝 Restart instance if memory usage is too high.
In Python 2, the memory check was always after the request returned,
so the request would still return HTTP 200 even if App Engine terminated
the instance for going over the memory limit.

In Python 3, there are additional memory checks mid-request. So
the same requests might return 500, 502, or 503 instead. We want to
pre-empt these errors by gracefully killing the instance ourselves
in advance.

We're finding that we use significantly more memory in Python 3
and will also need to investigate further why we use so much.

Bug: monorail:5271
Change-Id: I4646ac423093dec0861b54091690cbe2e07cd9ca