[UFS] Add indexer for Asset table
Bug: b/257084387

Adds an indexer job for the asset table. This job is meant to be run
using shivas and not meant to be run by the cron(hence the 200 Years).
Will inform oncall before running the job and will prob run it
early morning/late night MTV time so that we can be sure of not
blocking any traffic. It would also give us time to handle any issues
this causes. Do not trigger at the end of the week...

Change-Id: I4bca7ed12c21ee22b04b6af42e9deab7edd5c5ee