Export file definitions as part of dirmd bq export
Ensure that file metadata is exported as part of bq export.

DirBQRows is expanded to have a files column, which should follow
the same rules as dir except pointing to a path. Any file overrides
defined in mapping are exported to the BQ table.

Bug: 1431748
Change-Id: I3a0008f42a39ec9fe4c0b8582e2940d459bf13db
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/4471188
Reviewed-by: Chan Li <chanli@chromium.org>
Commit-Queue: Jeff Yoon <jeffyoon@google.com>
Cr-Commit-Position: refs/heads/main@{#56442}