Add no-op device config client
Skeleton to show approach. The DeviceConfigClient is intended to be the
entrypoint for all DeviceConfig functions (which will also be removed
from CrosInventoryClient, and use cases migrated).

This sets up the interface factory, interfaces, and a no-op client that
we will populate with the implementation in future CLs.

Note the parameters/return values for this client are expressed with the
`ufsdevice` proto. The previous client used invV2Api types, so there
will need to be some changes to the types passed into these calls

BUG=b/276898706
Test Plan=make test

Change-Id: Ic80c7e4a539a112f21c761619eaaef245b7aea9e
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/4500120
Commit-Queue: Eli Trexler <elijahtrexler@google.com>
Reviewed-by: Justin Suen <justinsuen@google.com>
Cr-Commit-Position: refs/heads/main@{#56491}