[golangbuild] implement build result sharing
This change implements build result sharing by having golangbuild check
datastore for a CAS digest. If it finds one for the matching build, it
will use that digest to try and fetch a prebuilt toolchain from CAS.

If it succeeds (it may fail because CAS resources have a TTL), then it
moves forward with that toolchain. If it fails, then it builds a Go
toolchain from scratch as usual. Once it finishes with the build, it
uploads the toolchain to CAS and updates the datastore with the new CAS
digest for the build.

Currently the matching criterion is that the string
$GOOS-$GOARCH-$revision must match. $GOOS and $GOARCH are always the
host's $GOOS and $GOARCH. $revision is a main repo commit hash, or a
main repo change number + patchset ($num-$patchset).

This change puts build result sharing behind an experiment to start
with. It also adds support and documentation for experiments.

Change-Id: I24a6469bb93c67021254d63d922576cdb952020e
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/4497226
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>
Reviewed-by: Heschi Kreinick <heschi@google.com>
Commit-Queue: Michael Knyszek <mknyszek@google.com>
Cr-Commit-Position: refs/heads/main@{#56569}