Remove unused modules in test-resources backend
A lot of these modules were added in an attempt to get rpcexplorer
working but aren't actually needed. Some of these might be needed when
we implement a sign in outside of rpcexplorer but we're not using them
yet. Also set up the routing to enable rpcexplorer

Bug: 1425974
Change-Id: Icf3cf155c7a3e6aae27cfc4027e65fd3af9794de
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/4506796
Commit-Queue: Struan Shrimpton <sshrimp@google.com>
Reviewed-by: Riley Wong <rgw@google.com>
Cr-Commit-Position: refs/heads/main@{#56572}