[WIB] Fix for missing cipd filenames in some cases
The cipd cache returned the pinned src as is. This causes problems
because cipd_src may or may not have a filename specified. This fixes
the bug and will only download/pin the cipd src once (irrespective
of the filename) and will return the correct src.

Change-Id: Idb8ba7382f89eaa776fd13b17d135a695cfb8859
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/4519188
Reviewed-by: Adam Todd <actodd@google.com>
Commit-Queue: Anushruth Sasivehalli <anushruth@chromium.org>
Cr-Commit-Position: refs/heads/main@{#56599}