[golangbuild] add structured all.bash output for majority of Go tests
Use a design that is both simple and forward- and backward-compatible.
This means it works on all supported release branches without needing
us to maintain (and verify) special cases for different major versions
of Go. It prioritizes ease of maintenance and confidence in correctness.

We can still start special-casing newer Go versions if that brings us
enough benefits to make it worthwhile. We can also adjust where the
split happens as more progress is made, or simply wait until all of
dist fully supports -json (very close now!) and major Go releases
that don't become unsupported.

This adds a little (justified) complexity but doesn't get in the way
of test sharding. In fact, we can decide not to shard the small subset
of dist tests, and only shard the 'go test std cmd' part. Or both if
really needed.

Also add a new LongTest input property, to be set on builders with
the "longtest" run mod. It seems better to centralize handling of
longtest mode in golangbuild, just like it's already done for race
mode. The builder configurations will set this property and not try
to handle longtest via env vars (something that doesn't work for
golang.org/x repo tests that need a -short flag instead), staying
responsible only for timeout scaling.

For go.dev/issue/59990.

Change-Id: I77e42e17487591d034724462a66fc108b8b27734
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/4518341
Reviewed-by: Michael Knyszek <mknyszek@google.com>
Reviewed-by: Heschi Kreinick <heschi@google.com>
Commit-Queue: Dmitri Shuralyov <dmitshur@google.com>
Cr-Commit-Position: refs/heads/main@{#56615}