[golangbuild] gate 'std cmd' test ResultDB uploading behind experiment
Trying this out found that one of the test cases in the standard library
is producing a test ID that doesn't match the ResultDB required pattern
of '^[[:print:]]{1,512}$', though the error doesn't say what it is.

Gate uploading to ResultDB behind an experiment while I sort this out.

Change-Id: I0aebc0df664e96e01b30cddb046ecc290eaefb73
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/4520426
Reviewed-by: Heschi Kreinick <heschi@google.com>
Commit-Queue: Dmitri Shuralyov <dmitshur@google.com>
Cr-Commit-Position: refs/heads/main@{#56620}