Skip cipd boostrapping packages on linux-arm64
https://ci.chromium.org/p/infra-internal/builders/prod/swarm-docker-arm64-image-builder

That builds docker images for linux-arm bots, and so also needs to run
on a linux-arm bot. Right now it's failing with:
 Errors:
   failed to resolve infra/tools/cloudbuildhelper/linux-arm64@git_revision:7088a6d730d06d978dfa0e9c5dd07e51aeb5916e (line 13): no such tag
   failed to resolve infra/tools/luci/logdog/logdog/linux-arm64@git_revision:fe9985447e6b95f4907774f05e9774f031700775 (line 14): no such tag

It likely broke after https://crrev.com/c/4416084. So this just skips
the hooks that were added in that CL when on linux-arm.

The builder really only needs just an infra/infra.git checkout, and
doesn't need any of the peripheral stuff. So if this doesn't work,
we could probably just skip bot_update all together and replace it
with a `git clone`.

Bug: None
Change-Id: I84371c440b89b56e9abc4af9ccf93860d492529b
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/4546124
Commit-Queue: Ben Pastene <bpastene@chromium.org>
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>
Cr-Commit-Position: refs/heads/main@{#56849}