[golangbuild] swap order of 'go test std cmd' and 'dist test'
The order is fairly arbitrary without a definitive winner (and would
not matter at all after sharding is added and they'd run in parallel),
but doing the larger standard library tests first may feel more natural.

A part of the motivation for the previous order was that it made the
'##### Test execution environment' all.bash header visible at the start
of the build rather than mid-way. That header is going away as part of
the switch to JSON output and will be replaced by a solution built into
LUCI (see crbug.com/1446433).

So let's give this order a try too.

Change-Id: Ib571f5603a42a01867d48ea94ccca0dbcd4cb9b6
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/4549678
Reviewed-by: Heschi Kreinick <heschi@google.com>
Reviewed-by: Michael Knyszek <mknyszek@google.com>
Commit-Queue: Dmitri Shuralyov <dmitshur@google.com>
Cr-Commit-Position: refs/heads/main@{#56883}