Add a daily file roll up table and fetch metric implementation
Also include the SQL to generate the materialized table to get it
under source control. This can definitely be improved. Since we only
support getting files for a given parent, the roll ups for
directories should really be cached instead of figured out each time.
We can't so this with a materialized table however, so that will
come in a follow up CL

Bug: 1435094
Change-Id: I45d3eba37a824fa1b61e165b04e6adce2db87e99