Add a daily file roll up table and fetch metric implementation
Also include the SQL to generate the table to get it under source
control. Unfortunately a materialized view won't work well since we
need to roll up by the sub directory and don't want to have to do
that with every fetch just to use the materialized view. Instead
we'll just do what we're doing with the weekly table update with a
normal table that gets updated at the same time.

Bug: 1435094
Change-Id: I45d3eba37a824fa1b61e165b04e6adce2db87e99
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/4545752
Commit-Queue: Struan Shrimpton <sshrimp@google.com>
Reviewed-by: Riley Wong <rgw@google.com>
Cr-Commit-Position: refs/heads/main@{#56914}