fix(DutTopology): move caching server state filtering to higher level
There are two consumer of the fetched caching server list: 1) subnet
based locator; 2) zone based locator. We need to filter the server
list by the state, but we only did it for subnet locator and forgot
to do the same thing for the newly added zone locator, which caused
a bug.

This CL moves the filtering from subnet locator to where we fetch the
list to avoid the bug.

BUG=b:283972253
TEST=run labservice locally and ran below RPC request
    echo '{"id":{"value": "<dut-name>"}}'|~/go/bin/grpc-client-cli \
    -s chromiumos.test.lab.api -m GetDutTopology  \
    --proto <path>/src/config/proto/chromiumos/test/lab/api/inventory_service.proto \
    --protoimports <path>/src/config/proto/ \
    localhost:1485

Change-Id: Ib40c14b30c61825aaebfca8316e1666743c5c0a7
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/4555890
Auto-Submit: Congbin Guo <guocb@chromium.org>
Reviewed-by: Vineet Joshi <vkjoshi@google.com>
Commit-Queue: Vineet Joshi <vkjoshi@google.com>
Cr-Commit-Position: refs/heads/main@{#56951}