Support patterned histograms in metrics Tricium analyzer
Analyze patterned histograms in Tricium. This includes <variants> defined out-of-line and <variant> defined inside <token> tags. It doesn't work for <variants> defined in other files, which shouldn't be an issue with the new histogram guidelines and the cleanup in progress.

Also apply some fixes to ensure removing an already deprecated or obsoleted histogram won't show a Tricium info suggesting to add an obsoletion message.

Limitations:
1. Only histograms.xml files are analyzed for histogram removals, so this check doesn't work well with suffix histograms.
2. File-level Tricium message will include histograms moved to a different file.

Bug: 1445725
Bug: b/214597409
Change-Id: I48685ac3f3e76db80991070666721ba1fdbb74f5