ufs: Add GetMachineACL
When we retrieve from datastore, use the realm on the entity and check
if user can access. We have to use the entity's realm (and not the
proto's realm) since there is no guarantee that the proto has a realm.

I couldn't find a super clean implementation here. Since golang isn't
covariant, I couldn't use the exact same function for `ufsds.Get` and
`ufsds.GetACL`, even though in practice `MachineEntity` would be both a
FleetEntity and RealmEntity. So I think it makes sense to just have two
functions, with one of them simply calling the other one.

Finally, I decided to throw the ACL logic into `registration`, rather
than adding it to all of the places that call `GetMachine`. This is
because we would have to duplicate the logic many times. Additionally,
this lets us clearly separate cases where we want to roll out ACLs, vs
un-ACLed uses of `GetMachine`

BUG=b/257084387
Test Plan=go test

Change-Id: Id9951636828c1614760ad82541937ea1c4499bef
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/4574773
Commit-Queue: Eli Trexler <elijahtrexler@google.com>
Reviewed-by: Anushruth Sasivehalli <anushruth@chromium.org>
Cr-Commit-Position: refs/heads/main@{#57148}