ufs: Use GetMachineACL
Replaces most, but not all use cases of GetMachine with GetMachineACL.
Does not change automated use cases.

The underlying `GetMachineACL` distinguishes between `NotFound` and
`PermissionDenied`, which we can take advantage of to use this method
broadly. For example, when updating an asset, we check for a
corresponding machine. If it is not found, we create the machine.
However, in the case where a machine exists, but the user doesn't have
permission to see it, we avoid creating a new machine.

This lets us use GetMachineACL more broadly which I think is useful.

BUG=b/257084387
Test Plan=make test, manual testing
Using an account with permissions in only one realm, went through all
use cases we have used GetMachineACL in, verifying positive and negative
cases

Change-Id: I10d9121154b28b02bcebf98273aa71714c3873fc
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/4574774
Commit-Queue: Eli Trexler <elijahtrexler@google.com>
Reviewed-by: Anushruth Sasivehalli <anushruth@chromium.org>
Cr-Commit-Position: refs/heads/main@{#57171}