siso: don't forget depfile
for deps=gcc, we don't need to keep it in hashfs, because
deps_gcc will parse it, stores info in deps log and remove depsfile.

but for not deps=gcc, we need to keep it in hashfs as it may be
used as outputs and should have cmdhash to check whether it should
need to run or not in incremental build.

Flush to make sure mtime matches between hashfs and local disk,
after mtime is updated to the time when step ran.

Bug: b/254182269
Change-Id: I70d7f30d8f61f13582dee4229d3a36e2b9389d25
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/4584499
Auto-Submit: Fumitoshi Ukai <ukai@google.com>
Reviewed-by: Takuto Ikuta <tikuta@chromium.org>
Commit-Queue: Fumitoshi Ukai <ukai@google.com>
Cr-Commit-Position: refs/heads/main@{#57181}