ufs: Make ufsds.ExistsACL method
This returns a boolean array where the i-th element represents whether
or not the i-th entity exists AND is visible to the user.

This uses a GetMulti behind the scenes which the ufsds.Exists() method
also uses, so efficiency should not be a huge concern.

Test Plan=make test
BUG=b/285604480

Change-Id: I95f47588f6096d7aab7eafe337ed42c3cb868e47
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/4594124
Commit-Queue: Eli Trexler <elijahtrexler@google.com>
Reviewed-by: Anushruth Sasivehalli <anushruth@chromium.org>
Cr-Commit-Position: refs/heads/main@{#57317}