[golangbuild] document cloud-platform scope, annotate errors
The cloud-platform scope is now needed for sourceForBranch to be able to
use a Gitiles REST client. Without it, trying a local build and run will
fail inscrutably with:

	set status: INFRA_FAILURE: interactive login is required
	exit status 1

Add the scope and annotate more errors. Not all are actually needed for
this specifically, but since I already have them I think adding more is
probably harmless. (I think it's fine to defer adding these until there
is a concrete need such as this one.)

Change-Id: Ide372ac3fdd5d9fca36174f50b14ca90c0cc6323
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/4598617
Reviewed-by: Michael Knyszek <mknyszek@google.com>
Commit-Queue: Dmitri Shuralyov <dmitshur@google.com>
Cr-Commit-Position: refs/heads/main@{#57319}