[golangbuild] move nested modules to independent directories
At this time, the go command doesn't have a direct way of catching tests
that read testdata across a module boundary. Such tests will fail when
the nearby modules aren't in the same relative position as they are in
the repository holding them.

Ideally the go command may in the future make this easier to detect for
Go developers running tests in their local repository checkouts, but we
take on a little complexity to help catch such problems in golangbuild
instead for now.

Move each nested module into a separate directory, with a pseudo-random
path, to detect and fix the tests that would otherwise only be detected
and reported by users much later.

x/tools, x/telemetry, and x/exp are known to need more attention before
being able to take advantage of this, are left out of scope of this CL.
Trying this on other repos sooner may help find new unknowns.

For go.dev/issue/34352.
For go.dev/issue/60666.

Change-Id: Ieac4c28d1920980393818178199065649691c860
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/4598622
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>
Reviewed-by: Michael Knyszek <mknyszek@google.com>
Commit-Queue: Dmitri Shuralyov <dmitshur@google.com>
Cr-Commit-Position: refs/heads/main@{#57328}