ufs: Forward user creds to inventory
Currently, UFS uses it's own service account when making requests to
inventory. This is a blocker to go/sfp-mvp-deviceconfig, as it means
partners who have access to UFS (expected), will have RPCs to inventory
service authenticated (unexpected), giving them information they may not
be entitled to.

This forwards the user's credentials to inventory instead.

Tested via calling `./dev-shivas internal-print-bot-info` and using the
UFS logics to validate the caller identify was being forwarded
correctly, and cros-inventory-dev was accepting/rejecting the request
appropriately with:
- my @google.com account (able to call inventory)
- my partner account (rejected inventory call)
- skylab-drone SA (able to call inventory)

It is possible there is some user of UFS who does not have direct
inventory access that this CL will break their access. However, if that
is the case, we can just add them to the
`chromeos-inventory-readonly-access` CRIA group to grant them access.

BUG=b/285604480
Test Plan=see commit

Change-Id: I124e6bfed61fc2fba4a1eca13c4f682ab0cdf9f5
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/4594163
Commit-Queue: Eli Trexler <elijahtrexler@google.com>
Reviewed-by: Justin Suen <justinsuen@google.com>
Cr-Commit-Position: refs/heads/main@{#57352}