[UFS] Machine: Enforce realm assignment based on zone
Bug: b/285603336

Even though the realm is assigned using zone. The assignment itself was
left to clients. This means that shivas is the only client that can be
trusted to assign the realms reliably. As we should not trust any clients
even shivas for enforcing our ACLs. This change makes realm a read only
field and explictly assigns realm to the proto and entity. This means it
also needs to fix the unit tests that were dependent on realm assignment.
The following changes were made to accomodate the new realm assignment
behavior

  1.  Add fake user and permissions for the cases where it didn't exist
  2.  Replace realm assignment with one of the correct location/zone
      assignments. For example ATL97 is one of the zones in browser
      realm.

      Realm: util.BrowserLabAdminRealm

      changes to

      Location: &ufspb.Location{
        Zone: ufspb.Zone_ZONE_ATL97
      }
  3.  Add withAuthorized{Atl|Acs}User helpers for assigning mock
      permissions for unit tests.
  4.  Use a default user and permissions for all the frontend tests.
      (Not expected to test ACLs in frontend unit tests)

Change-Id: Ic71a9c22d0612a4ecf939bafa15210573ef7c7d8
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/4599829
Reviewed-by: Gregory Nisbet <gregorynisbet@google.com>
Commit-Queue: Anushruth Sasivehalli <anushruth@chromium.org>
Cr-Commit-Position: refs/heads/main@{#57371}