[UFS] MachineLSE: Add support to index MachineLSE table
Bug: b/285603337

This patch if applied adds support to index the complete MachineLSE
table with the correct ACLs. This indexing is meant to be run from
shivas and will not, under normal circumstances be triggered. The run
time for this job will be significantly longer than other reindexing
jobs because of the multiple reads/writes involved.

Ran the cron job with local instance of UFS and writes disabled. The
dev instances entries took about 10 minutes to list (Nothing was
updated). Expecting it to take atleast an hour to update everything
on prod.

Change-Id: I4563ecfa05581d1570597aa631bcfef1fd740b60
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/4603337
Reviewed-by: Eli Trexler <elijahtrexler@google.com>
Commit-Queue: Anushruth Sasivehalli <anushruth@chromium.org>
Reviewed-by: Gowri Denduluri <gowriden@google.com>
Cr-Commit-Position: refs/heads/main@{#57385}