Support CL-level obsoletion message
Tricium will post a comment containing the list of histograms removed in the CL to the commit message. The exception is suffix histograms, only the base histogram will be recorded.

Stop suggesting to add an obsoletion message for specific histograms if there's a CL-level obsoletion message present.

Histograms that are obsolete or deprecated will be ignored.

Draft of the corresponding server-side change can be found in cl/535439181.

Bug: b/214597409
Change-Id: Iaaf0eaa411fe521575cdbce9586d373e5194436a
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/4573861
Reviewed-by: Jesse Doherty <jwd@chromium.org>
Reviewed-by: Yiwei Zhang <yiwzhang@google.com>
Cr-Commit-Position: refs/heads/main@{#57390}