Add fallback config repo paths for dependency projects.
If the checkout location of the dependency repo is moved, it's possible
that a CL could be sent that is based on a revision from before the new
path was added. In that case, if the bootstrapper config has been
updated to only look for the new path, then the bootstrapper will error
out when it tries to get the pinned revision from the CL and its base.
This adds the fallback_config_repo_paths field to DependencyProject,
which will allow safely switching the bootstrapper to read from the new
location without erroring out on CLs that have an older base revision.

Bug: 1454296
Change-Id: I65840fe4aa632242861d9c9ebf87aba649e09409
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/4611047
Auto-Submit: Garrett Beaty <gbeaty@google.com>
Reviewed-by: Struan Shrimpton <sshrimp@google.com>
Commit-Queue: Struan Shrimpton <sshrimp@google.com>
Cr-Commit-Position: refs/heads/main@{#57411}