ufs: Skip HWID in partner namespace
This will allow us to see the impact of not assigning HWID data
dimensions to the swarming bots, and let us make a decision about
how urgent supporting HWID is for Satlab for Partners.

If this is the strategy for MVP, we will also need to enforce namespace
restrictions for partners (ie only let them interact with `os-partner`
namespace). Until now this has not been a security boundary but an
organizational one.

Long term, we will want to replace this with the real HWID service: b/286921043

BUG=b/285605479
Test Plan=`make test`. Also `make dev`, then ran `./dev-shivas
internal-print-bot-info` against the os-partner and os namespaces. In os
namespace, used logs to verify we hit HWID service. In os-partner
namespace, used logs to verify we did not hit HWID service

Change-Id: I9536c4b72ec1c754cd48e27cb8ed5903e575de61
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/4608831
Commit-Queue: Eli Trexler <elijahtrexler@google.com>
Reviewed-by: Justin Suen <justinsuen@google.com>
Cr-Commit-Position: refs/heads/main@{#57430}